pyproject.toml
[project]
name = "multi-turn-dialogue-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "John",email = "jianghao36@163.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.2,<0.7.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-qianwen (>=0.1.17,<0.2.0)",
    "langchain-tavily (>=0.2.11,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
